Initial gui refactor, divided gui file into file for main screen and parameter screen, created gui utilities to avoid duplicating code
