+++ src/main.py
@@ -1,7 +1,7 @@
+from gui import gui_run
 from logger import init_logger
-import gui
 
 if __name__ == "__main__":
     init_logger()
-    gui.run()
+    gui_run.run()
